Rename group_store.rs to store.rs
