+++ p2panda-auth/src/traits/mod.rs
@@ -4,14 +4,14 @@ use std::fmt::Debug;
 use std::hash::Hash as StdHash;
 
 mod dgm;
-mod group_store;
+mod store;
 mod operation;
 mod ordering;
 mod query;
 mod resolver;
 
 pub use dgm::GroupMembership;
-pub use group_store::GroupStore;
+pub use store::GroupStore;
 pub use operation::Operation;
 pub use ordering::Ordering;
 pub use query::GroupMembershipQuery;
