+++ p2panda-auth/src/traits/store.rs
